backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
bcrypt==4.2.0
argon2-cffi==23.1.0
passlib==1.7.4
python-multipart==0.0.9
